Testing

+++ contracts/interface/INestLedger.sol
@@ -9,10 +9,10 @@ interface INestLedger {
     struct Config {
         
         // nest reward scale(10000 based). 2000
-        uint32 nestRewardScale;
+        uint16 nestRewardScale;
 
         // ntoken reward scale(10000 based). 8000
-        uint32 ntokenRedardScale;
+        uint16 ntokenRedardScale;
     }
     
     /// @dev Modify configuration
